% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Load data into project}
\usage{
bind_data(project, site_data, genetic_data)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{site_data}{TODO}

\item{genetic_data}{TODO}
}
\description{
TODO
}
