% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_hex_grid}
\alias{create_hex_grid}
\title{Create map composed of hexagonal tiles}
\usage{
create_hex_grid(project, hex_width = NULL, border_coords = NULL)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{hex_width}{width of hexagons.}

\item{border_coords}{dataframe giving coordinates (longitude, latitude) of a
polygon within which the map is defined. If null then this is generated
automatically from the convex hull of the sample site locations.}
}
\description{
TODO
}
