% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_model}
\alias{define_model}
\title{Define parameters of the spatial correlation model}
\usage{
define_model(project, alpha_0 = 1, beta_0 = 1, phi_0 = 0, gamma_0 = 0.1)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{alpha_0}{TODO}

\item{beta_0}{TODO}

\item{phi_0}{TODO}

\item{gamma_0}{TODO}
}
\description{
TODO
}
