% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GeoMAPI_analysis}
\alias{GeoMAPI_analysis}
\title{TODO}
\usage{
GeoMAPI_analysis(
  project,
  reps = 2,
  inner_reps = 10,
  silent = FALSE,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{reps}{TODO}

\item{inner_reps}{TODO}

\item{silent}{TODO}

\item{pb_markdown}{TODO}
}
\description{
TODO
}
