% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GeoMAPI_assign_edges}
\alias{GeoMAPI_assign_edges}
\title{Assign edges for GeoMAPI analysis}
\usage{
GeoMAPI_assign_edges(
  project,
  eccentricity = 0.9,
  max_dist = Inf,
  silent = FALSE,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{eccentricity}{eccentricity of ellipses, defined as half the distance
between foci divided by the semi-major axis. \eqn{e = sqrt{1 - b^2/a^2}},
where \eqn{e} is the eccentricity, \eqn{a} is the length of the semi-major
axis, and \eqn{b} is the length of the semi-minor axis. Eccentricity ranges
between 0 (perfect circle) and 1 (straight line between foci). An
eccentricity of 0 is not allowed in this case because it would result in an
infinitely large circle.}

\item{max_dist}{edges shorter than this length are discarded prior to
assignment. Can be used to focus on short distance signals.}

\item{silent}{if \code{TRUE} then no output is produced during function
evaluation.}

\item{pb_markdown}{whether to run progress bars in markdown mode, in which
case they are updated once at the end to avoid large amounts of output.}
}
\description{
Given an GeneScapeR project with a grid already loaded (see
  \code{?create_hex_grid}), determine which pairwise edges in the data
  intersect each grid cell. Assumes an elliptical projection along each edge
  with the start- and end-points becoming the two foci of the ellipse, and
  with the eccentricity defined by the user. See Details for suggestions on
  how to choose this parameter.
}
\details{
TODO
}
