% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GeoMAPI_get_significant}
\alias{GeoMAPI_get_significant}
\title{Get GeoMAPI significant cells}
\usage{
GeoMAPI_get_significant(
  project,
  test_tail = "both",
  FDR = 0.05,
  min_coverage = 10
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{test_tail}{whether to calculate empirical p-values using a
one-sided test (\code{test_tail = "left"} or \code{test_tail =
"right"}) or a two-sided test (\code{test_tail = "both"}).}

\item{FDR}{the false discovery rate, i.e. the probability that a cell
identified as significant is actually a false positive.}

\item{min_coverage}{minimum coverage (number of edges assigned to a cell)
for it to be included in the final result.}
}
\description{
Given a completed GeoMAPI analysis, identify which cells are
  significant outliers.
}
