% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GeoMAPI_plot_coverage}
\alias{GeoMAPI_plot_coverage}
\title{Plot GeoMAPI coverage}
\usage{
GeoMAPI_plot_coverage(
  project,
  breaks = c(0, 10, 20, 30, 40, 50, 100, Inf),
  col_scale = col_hotcold,
  plot_sampling_points = TRUE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{breaks}{the sequence of coverage breaks used.}

\item{col_scale}{the colour scale to use.}

\item{plot_sampling_points}{whether to overlay sampling locations.}
}
\description{
TODO
}
