% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GeoMAPI_plot_zscore}
\alias{GeoMAPI_plot_zscore}
\title{Plot GeoMAPI z-scores and significance}
\usage{
GeoMAPI_plot_zscore(
  project,
  plot_sampling_points = TRUE,
  plot_significance = TRUE,
  min_hex_coverage = 10,
  col_scale = viridisLite::viridis,
  zlim = NULL,
  base_plot = NULL,
  point_size = 1,
  point_colour = "white",
  point_fill = "black",
  point_stroke = 0.2
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{plot_sampling_points}{whether to overlay sampling locations.}

\item{plot_significance}{whether to outline areas that were identified as
significant outliers.}

\item{min_hex_coverage}{minimum coverage (number of edges assigned to a hex)
for it to be included in the final result, otherwise these hexes are given
the value \code{NA}.}

\item{col_scale}{the colour scale to use.}

\item{zlim}{the limits of the colour scale. If \code{NULL} then these limits
are chosen automatically.}

\item{base_plot}{optional base plot (object of class \code{ggplot}) on which
this function builds. If \code{NULL} then a simple empty plot is used.}

\item{point_size, point_colour, point_fill, point_stroke}{properties of plotted
sampling points.}
}
\description{
TODO
}
