% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GeoMAPI_suggest_eccentricity}
\alias{GeoMAPI_suggest_eccentricity}
\title{Suggest eccentricity parameter based on target coverage}
\usage{
GeoMAPI_suggest_eccentricity(
  project,
  target_coverage = 10,
  target_proportion = 0.9,
  n_iterations = 30,
  max_dist = Inf,
  silent = FALSE,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{target_coverage, target_proportion}{eccentricity will be chosen such
that a proportion \code{target_proportion} of cells achieve a coverage of
at least \code{target_coverage}.}

\item{n_iterations}{the number of iterations to run of the binary search
method.}

\item{max_dist}{edges shorter than this length are discarded prior to
assignment. Can be used to focus on short distance signals.}

\item{silent}{if \code{TRUE} then no output is produced during function
evaluation.}

\item{pb_markdown}{whether to run progress bars in markdown mode, in which
case they are updated once at the end to avoid large amounts of output.}
}
\description{
TODO
}
\details{
TODO
}
