% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ellipse}
\alias{get_ellipse}
\title{Calculate ellipse polygon coordinates from foci and eccentricity}
\usage{
get_ellipse(f1 = c(-3, 0), f2 = c(3, 0), ecc = 0.8, n = 100)
}
\arguments{
\item{f1}{x- and y-coordinates of the first focus.}

\item{f2}{x- and y-coordinates of the second focus.}

\item{ecc}{eccentricity of the ellipse, defined as half the distance between
foci divided by the semi-major axis. We can say \eqn{e = sqrt{1 -
b^2/a^2}}, where \eqn{e} is the eccentricity, \eqn{a} is the length of the
semi-major axis, and \eqn{b} is the length of the semi-minor axis.
Eccentricity ranges from 0 (perfect circle) to 1 (straight line between
foci), although eccentricity of 0 is not allowed in this function as it
would lead to an infinitely large circle.}

\item{n}{number of points in polygon.}
}
\description{
TODO calculate ellipse polygon coordinates from foci and eccentricity.
}
