% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_GC_distance}
\alias{get_GC_distance}
\title{Calculate great circle distance between pairwise coordinates}
\usage{
get_GC_distance(lon, lat)
}
\arguments{
\item{lon, lat}{vector of longitudes and latitudes.}
}
\description{
Get great circle distance between spatial points defined by a
  vector of longitudes and latitudes. Distances are returned in a pairwise
  distance matrix.
}
