% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_mu_sigsq_credible}
\alias{get_mu_sigsq_credible}
\title{TODO}
\usage{
get_mu_sigsq_credible(
  project,
  loci,
  reps = 10,
  quantiles = c(0.025, 0.5, 0.975),
  silent = FALSE,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{TODO}

\item{loci}{TODO}

\item{reps}{TODO}

\item{quantiles}{TODO}

\item{silent}{TODO}

\item{pb_markdown}{TODO}
}
\description{
TODO
}
