% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_pairwise_D}
\alias{get_pairwise_D}
\title{Calculate pairwise Jost's D between sites}
\usage{
get_pairwise_D(project)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}
}
\description{
TODO
}
