% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairwise}
\alias{plot_pairwise}
\title{Plot pairwise spatial distance vs. genetic distances}
\usage{
plot_pairwise(project, genetic_stat = "Gst")
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{genetic_stat}{TODO}
}
\description{
TODO
}
