% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior_cor}
\alias{plot_prior_cor}
\title{Produce visualisation of prior on spatial correlation}
\usage{
plot_prior_cor(project, type = 1, n_reps = 10, d_max = NULL)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{type}{TODO}

\item{n_reps}{TODO}

\item{d_max}{TODO}
}
\description{
TODO
}
