% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior_freq}
\alias{plot_prior_freq}
\title{Produce visualisation of prior on allele frequencies}
\usage{
plot_prior_freq(project, n_reps = 10, p_step = 0.01)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{n_reps}{TODO}

\item{p_step}{TODO}
}
\description{
TODO
}
