% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spatial_autocor}
\alias{plot_spatial_autocor}
\title{Plot prior and/or posterior spatial correlation}
\usage{
plot_spatial_autocor(
  project,
  prior_on = TRUE,
  posterior_on = TRUE,
  n_reps = 1000,
  d_max = NULL
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{prior_on}{TODO}

\item{posterior_on}{TODO}

\item{n_reps}{TODO}

\item{d_max}{TODO}
}
\description{
TODO
}
