% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_map}
\alias{predict_map}
\title{TODO}
\usage{
predict_map(
  project,
  loci,
  reps = 2,
  inner_reps = 10,
  quantiles = c(0.025, 0.5, 0.975),
  exceedance = NULL,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{loci}{TODO}

\item{reps}{TODO}

\item{inner_reps}{TODO}

\item{quantiles}{TODO}

\item{exceedance}{TODO}

\item{pb_markdown}{TODO}
}
\description{
TODO
}
