% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_pairwise}
\alias{predict_pairwise}
\title{TODO}
\usage{
predict_pairwise(
  project,
  reps = 2,
  inner_reps = 10,
  quantiles = c(0.025, 0.5, 0.975),
  silent = FALSE,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{reps}{TODO}

\item{inner_reps}{TODO}

\item{quantiles}{TODO}

\item{silent}{TODO}

\item{pb_markdown}{TODO}
}
\description{
TODO
}
