% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC}
\usage{
run_mcmc(project, chains = 1, init_list = NULL, ...)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{chains}{TODO}

\item{init_list}{TODO}

\item{...}{additional parameters that will be passed to
\code{drjacoby::run_mcmc()}.}
}
\description{
TODO
}
