% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_GRF}
\alias{sim_GRF}
\title{TODO}
\usage{
sim_GRF(
  project,
  loci,
  alleles,
  nu,
  lambda,
  omega,
  sigsq_shape,
  sigsq_rate,
  mu_mean,
  gamma
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function. The project must have a grid already
defined (see \code{?create_hex_grid}) as this grid will be used in
simulation.}

\item{loci}{number of loci (assumed independent).}

\item{alleles}{number of alleles, assumed to be the same over all loci.}

\item{nu, lambda, omega}{the values of the model parameters used in simulation.
These parameters together specify the spatial autocorrelation function.}

\item{sigsq_shape, sigsq_rate, mu_mean, gamma}{parameters of the
normal-inverse-gamma prior on the mean and variance of transformed allele
frequencies.}
}
\description{
TODO
}
