% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Wombling}
\alias{Wombling}
\title{TODO}
\usage{
Wombling(
  project,
  loci = NULL,
  reps = 2,
  inner_reps = 10,
  measure = "all",
  patch_size = 1,
  quantiles = c(0.025, 0.5, 0.975),
  silent = FALSE,
  pb_markdown = FALSE
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{loci}{TODO}

\item{reps}{TODO}

\item{inner_reps}{TODO}

\item{measure}{TODO}

\item{patch_size}{TODO}

\item{quantiles}{TODO}

\item{silent}{TODO}

\item{pb_markdown}{TODO}
}
\description{
TODO
}
