% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Wombling_get_significant}
\alias{Wombling_get_significant}
\title{TODO}
\usage{
Wombling_get_significant(
  project,
  measure = "all",
  test_tail = "both",
  FDR = 0.05,
  silent = FALSE
)
}
\arguments{
\item{project}{TODO}

\item{measure}{TODO}

\item{test_tail}{TODO}

\item{FDR}{TODO}

\item{silent}{TODO}
}
\description{
TODO
}
