% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_model}
\alias{define_model}
\title{Define parameters of the spatial correlation model}
\usage{
define_model(
  project,
  nu_shape1 = 1,
  nu_shape2 = 1,
  lambda_shape = 1,
  lambda_rate = NULL
)
}
\arguments{
\item{project}{a genescaper project, as produced by the
\code{genescaper_project()} function.}

\item{nu_shape1}{TODO}

\item{nu_shape2}{TODO}

\item{lambda_shape}{TODO}

\item{lambda_rate}{TODO}
}
\description{
TODO
}
