% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_GRF}
\alias{sim_GRF}
\title{TODO}
\usage{
sim_GRF(
  project,
  loci,
  alleles,
  nu,
  lambda,
  omega,
  sigsq_shape,
  sigsq_rate,
  mu_mean,
  mu_scale
)
}
\arguments{
\item{N}{TODO}
}
\description{
TODO
}
